kafka-pipeline/requirements.txt
confluent-kafka==2.6.1
pydantic==2.9.2
httpx==0.27.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
